+++ src/main/java/com/bitbank/dto/v1/BankAccountTransactionsDTO.java
@@ -14,9 +14,5 @@ public class BankAccountTransactionsDTO {
 	
 	private String purpose;
 	
-	public Double getAmount() {
-		return this.amount.doubleValue();
-	}
-	
 	private LocalDateTime transactionDate;
 }