+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 + Add Bank Account Transactions controller + DTO
 + Add Docker for deploy
 + Add Bank Account Transactions service POST
++ Add Bank Account Transactions controller POST
 
 ## 0.1.0
 + First commit